+++ webpack.config.js
@@ -1,12 +1,12 @@
-const fs = require('fs');
-const path = require('path');
+const fs = require('node:fs');
+const path = require('node:path');
 const ESLintPlugin = require('eslint-webpack-plugin');
 const HtmlWebPackPlugin = require('html-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 require('dotenv-safe').config();
-const appUser = Buffer.from(process.env.APPUSER).toString('base64');
-const noAlien = Buffer.from(process.env.NOALIEN).toString('base64');
+const appUser = Buffer.from(process.env.APPUSER || '').toString('base64');
+const noAlien = Buffer.from(process.env.NOALIEN || '').toString('base64');
 
 const getPackageVersion = () => {
   let pkg = JSON.parse(fs.readFileSync('./package.json').toString());
@@ -73,7 +73,16 @@ module.exports = {
       },
       {
         test: /\.less$/,
-        use: [{ loader: MiniCssExtractPlugin.loader }, { loader: 'css-loader' }, { loader: 'less-loader' }]
+        use: [
+          { loader: MiniCssExtractPlugin.loader },
+          { loader: 'css-loader' },
+          {
+            loader: 'less-loader',
+            options: {
+              lessOptions: { javascriptEnabled: true }
+            }
+          }
+        ]
       }
     ]
   },